Refactor project structure, start implementing BEM methodology
Rework frontend architecture, initiating gradual adoption of BEM naming convention for improved code maintainability and scalability.
