+++ front-page.php
@@ -5,8 +5,6 @@ if (! defined('ABSPATH')) {
   exit;
 }
 get_header();
-
-get_template_part('template-parts/pagebuttons', 'basic');
 ?>
 <!--Content-->
 <div class="container">
