+++ resources/assets/sass/components/container/_bootstrap-container.scss
@@ -1,0 +1,8 @@
+// Configuration
+@import "../../../libs/bootstrap/scss/functions";
+@import "../../../libs/bootstrap/scss/variables";
+@import "../../../libs/bootstrap/scss/mixins";
+// @import "../../../libs/bootstrap/scss/utilities";
+
+//Layout & components
+@import "../../../libs/bootstrap/scss/containers";
