+++ resources/assets/sass/components/navbar/_navbar-mixins.scss
@@ -1,0 +1,37 @@
+@import '../../mixins/colors';
+
+@mixin navbar-light {
+    color: $dark-color;
+    background-color: $light-color;
+
+    a {
+        color: $dark-color;
+    }
+}
+
+@mixin navbar-dark {
+    color: $light-color;
+    background-color: $dark-color;
+
+    a {
+        color: $light-color;
+    }
+}
+
+@mixin navbar-transparent {
+    color: $light-color;
+    background-color: transparent;
+
+    a {
+        color: $light-color;
+    }
+}
+
+@mixin navbar-primary {
+    color: $dark-color;
+    background-color: $primary-color;
+
+    a {
+        color: $dark-color;
+    }
+}
