+++ resources/assets/sass/mixins/_colors.scss
@@ -1,0 +1,12 @@
+$primary-color: #007bff;
+$secondary-color: #6c757d;
+
+$light-color: #f8f9fa;
+$dark-color: #343a40;
+
+$danger-color: #dc3545;
+$success-color: #28a745;
+$warning-color: #ffc107;
+$info-color: #17a2b8;
+
+// ===================
