+++ resources/assets/sass/mixins/_flex.scss
@@ -1,0 +1,5 @@
+@mixin flex-justify($value) {
+    display: flex;
+    align-items: center;
+    justify-content: $value;
+  }
